+++ GeniusTests/Extensions/Genius/Symbol.swift
@@ -10,7 +10,7 @@ import SwiftUI
 
 extension Symbol: @retroactive Equatable {
 
-	static let notApplicable = Self(.minus, color: .primary, label: .notApplicable)
+	static let notAvailable = Self(.minus, color: .primary, label: .notAvailable)
 	static let disabled = Self(.xmark, color: .red, label: .disabled)
 	static let enabled = Self(.checkmark, color: .green, label: .enabled)
 	static let unknown = Self(.questionmark, color: .red, label: .unknown)